Move websocket function to client
