+++ core/endpoint.go
@@ -344,7 +344,7 @@ func (endpoint *Endpoint) call(result *Result) {
 	} else if endpointType == EndpointTypeICMP {
 		result.Connected, result.Duration = client.Ping(strings.TrimPrefix(endpoint.URL, "icmp://"), endpoint.ClientConfig)
 	} else if endpointType == EndpointTypeWS {
-		queryWebSocket(endpoint, result)
+		result.Connected, result.Body, err = client.queryWebSocket(endpoint.URL, endpoint.ClientConfig)
 		result.Duration = time.Since(startTime)
 	} else {
 		response, err = client.GetHTTPClient(endpoint.ClientConfig).Do(request)
